Turned `TracingMessagingMessageListenerAdapter` into a customsizable bean
